use 64bit value for the connection count

+++ trafalgar/tr_geo_object.h
@@ -184,7 +184,7 @@ public:
 
 	virtual uint8_t handleCrossing(const TrZoomMap & zoom_ref, TrGeoObject * other, TrGeoObject * node, uint8_t mode);
 
-	virtual uint32_t handleConnecion(TrGeoObject * node, bool dir, uint32_t flags);
+	virtual uint64_t handleConnecion(TrGeoObject * node, bool dir, uint64_t flags);
 
 	virtual void setType(uint16_t);
