+++ trafalgar/tr_map_link_road.h
@@ -157,7 +157,7 @@ public:
 
 	virtual uint8_t handleCrossing(const TrZoomMap & zoom_ref, TrGeoObject * other, TrGeoObject * node, uint8_t mode);
 
-	virtual uint32_t handleConnecion(TrGeoObject * node, bool dir, uint32_t flags);
+	virtual uint64_t handleConnecion(TrGeoObject * node, bool dir, uint64_t flags);
 
 #ifdef TR_SERIALIZATION
 	virtual uint64_t readXmlDescription(QXmlStreamReader & xml_in);
